Move `FeedImageStoreSpy` to `Helpers` group
